AArch64: Correction of definition of UINT32
In an #ifdef ENABLE_ANDROID_NULL_CONVERT...#endif section of the
jccolor.c module UINT32 was defined as unsigned long, which is OK
for 32-bit systems but incorrect for __LP64__ systems such as AArch64.
The definition of UINT32 has been changed to unsigned int, which is
32-bits on both 32-bit and __LP64__ systems.

+++ jccolor.c
@@ -302,7 +302,7 @@ grayscale_convert (j_compress_ptr cinfo,
 
 #ifdef ENABLE_ANDROID_NULL_CONVERT
 
-typedef unsigned long UINT32;
+typedef unsigned int UINT32;
 
 #define B0(n)   ((n) & 0xFF)
 #define B1(n)   (((n) >> 8) & 0xFF)
